requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --upgrade -o requirements/test.txt requirements/test.in
#
amqp==1.4.9               # via kombu
analytics-python==1.2.9
anyjson==0.3.3            # via kombu
asn1crypto==0.24.0        # via cryptography
astroid==1.5.3            # via pylint, pylint-celery
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
aws-sam-translator==1.11.0  # via cfn-lint
aws-xray-sdk==2.4.2       # via moto
billiard==3.3.0.23        # via celery
boto3==1.9.136            # via aws-sam-translator, moto
boto==2.49.0              # via moto
botocore==1.12.136        # via aws-xray-sdk, boto3, moto, s3transfer
celery==3.1.26.post2
certifi==2019.3.9         # via requests
cffi==1.12.3              # via cryptography
cfn-lint==0.19.1          # via moto
chardet==3.0.4            # via requests
click-log==0.3.2          # via edx-lint
click==7.0                # via click-log, code-annotations, edx-lint
code-annotations==0.3.1
coverage==4.5.3
cryptography==2.6.1       # via moto
ddt==1.2.1
defusedxml==0.6.0         # via python3-openid, social-auth-core
django-dynamic-fixture==2.0.0
django-extensions==2.1.6
django-guardian==1.5.0
django-model-utils==3.1.2
django-mysql==3.0.0.post1
django-simple-history==2.7.2
django-storages==1.7.1
django-user-tasks==0.1.5
django-waffle==0.16.0
django==1.11.20
djangorestframework-jwt==1.11.0  # via edx-drf-extensions
djangorestframework==3.9.2
docker-pycreds==0.4.0     # via docker
docker==3.7.2             # via moto
docutils==0.14            # via botocore
ecdsa==0.13.2             # via python-jose
edx-auth-backends==2.0.0
edx-django-release-util==0.3.1
edx-django-utils==1.0.3   # via edx-drf-extensions
edx-drf-extensions==2.2.0
edx-lint==1.1.2
edx-opaque-keys==0.4.4    # via edx-drf-extensions
edx-rest-api-client==1.9.2
factory-boy==2.11.1
faker==1.0.5
future==0.17.1            # via aws-xray-sdk, pyjwkest, python-jose
idna==2.8                 # via moto, requests
isort==4.3.17             # via pylint
jinja2==2.10.1            # via code-annotations, moto
jmespath==0.9.4           # via boto3, botocore
jsondiff==1.1.2           # via moto
jsonpatch==1.23           # via cfn-lint
jsonpickle==1.1           # via aws-xray-sdk
jsonpointer==2.0          # via jsonpatch
jsonschema==2.6.0         # via aws-sam-translator, cfn-lint
kombu==3.0.37             # via celery
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
mock==2.0.0
more-itertools==7.0.0     # via pytest
moto==1.3.8
newrelic==4.18.0.118      # via edx-django-utils
oauthlib==3.0.1           # via requests-oauthlib, social-auth-core
pathlib2==2.3.3           # via pytest
pathspec==0.5.9           # via yamllint
pbr==5.2.0                # via mock, stevedore
pluggy==0.9.0             # via pytest
psutil==1.2.1             # via edx-django-utils, edx-drf-extensions
py==1.8.0                 # via pytest
pyasn1==0.4.5             # via rsa
pycodestyle==2.5.0
pycparser==2.19           # via cffi
pycryptodomex==3.8.1      # via pyjwkest
pyjwkest==1.3.2           # via edx-drf-extensions, social-auth-core
pyjwt==1.7.1              # via djangorestframework-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.5  # via pylint-celery, pylint-django
pylint==1.7.6             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.8.0            # via edx-opaque-keys
pytest-cov==2.6.1
pytest-django==3.4.8
pytest==4.4.1             # via pytest-cov, pytest-django
python-dateutil==2.8.0    # via analytics-python, botocore, edx-drf-extensions, faker, moto
python-jose==3.0.1        # via moto
python-slugify==3.0.2     # via code-annotations
python3-openid==3.1.0     # via social-auth-core
pytz==2019.1
pyyaml==5.1               # via cfn-lint, code-annotations, edx-django-release-util, moto, yamllint
redis==2.8
requests-oauthlib==1.2.0  # via social-auth-core
requests==2.21.0          # via analytics-python, cfn-lint, docker, edx-drf-extensions, edx-rest-api-client, moto, pyjwkest, requests-oauthlib, responses, slumber, social-auth-core
responses==0.10.6
rest-condition==1.0.3     # via edx-drf-extensions
rsa==4.0                  # via python-jose
s3transfer==0.2.0         # via boto3
semantic-version==2.6.0   # via edx-drf-extensions
six==1.12.0               # via analytics-python, astroid, aws-sam-translator, cfn-lint, cryptography, django-dynamic-fixture, django-extensions, django-simple-history, docker, docker-pycreds, edx-auth-backends, edx-django-release-util, edx-drf-extensions, edx-lint, edx-opaque-keys, faker, mock, moto, pathlib2, pyjwkest, pylint, pytest, python-dateutil, python-jose, responses, social-auth-app-django, social-auth-core, stevedore, websocket-client
slumber==0.7.1            # via edx-rest-api-client
social-auth-app-django==1.2.0  # via edx-auth-backends
social-auth-core[openidconnect]==1.7.0  # via edx-auth-backends, social-auth-app-django
stevedore==1.30.1         # via code-annotations, edx-opaque-keys
text-unidecode==1.2       # via faker, python-slugify
urllib3==1.24.2           # via botocore, requests
websocket-client==0.56.0  # via docker
werkzeug==0.15.2          # via moto
wrapt==1.11.1             # via astroid, aws-xray-sdk
xmltodict==0.12.0         # via moto
yamllint==1.15.0
